Function bluenoisers::blue_noise_iter

pub fn blue_noise_iter(dimensions: Vec<f64>, min_distance: f64, k_abort: usize) -> BlueNoiseIterator

Creates an iterator over the blue noise samples, generating them on demand. This is useful for pipelined processing or when you only need to take some amount of samples Otherwise this is the same as blue_noise