Crate bluenoisers

bluenoisers

Implementation of blue noise in Rust using Fast Poisson Disk Sampling. This implementation can be used for images (2-dimensional), for volumes (3-dimensional) and in higher dimensions.

For background information see here: https://www.cs.ubc.ca/~rbridson/docs/bridson-siggraph07-poissondisk.pdf

Documentation is here: https://connyonny.github.io/bluenoisers/bluenoisers

Structs

BlueNoiseIterator
The iterator struct returned by blue_noise_iter

Functions

blue_noise
Generates blue noise samples

blue_noise_iter
Creates an iterator over the blue noise samples, generating them on demand. This is useful for pipelined processing or when you only need to take some amount of samples Otherwise this is the same as blue_noise